feat(router): React router added. Layout folder containing routes. Small stylings added to RecipeItem. RecipeInfo interface created
