+++ src/Layout/NotFound.tsx
@@ -1,0 +1,9 @@
+function NotFound() {
+    return (
+      <div className="NotFound">
+        <h1>Not Found</h1>
+      </div>
+    );
+  }
+  
+  export default NotFound
